Remove AccordionHeaderExpandIcon
