+++ packages/react-accordion/src/components/AccordionHeader/index.ts
@@ -1,5 +1,4 @@
 export * from './AccordionHeader';
-export * from './AccordionHeaderExpandIcon';
 export * from './AccordionHeader.types';
 export * from './renderAccordionHeader';
 export * from './useAccordionHeader';
