+++ packages/react-components/src/index.ts
@@ -65,7 +65,6 @@ export {
   Accordion,
   AccordionContext,
   AccordionHeader,
-  AccordionHeaderExpandIcon,
   AccordionItem,
   AccordionItemContext,
   AccordionPanel,
@@ -102,7 +101,6 @@ export type {
   AccordionHeaderContextValue,
   AccordionHeaderContextValues,
   AccordionHeaderExpandIconPosition,
-  AccordionHeaderExpandIconProps,
   AccordionHeaderProps,
   AccordionHeaderSize,
   AccordionHeaderSlots,
